+++ tests/clippy/0005.t
@@ -54,7 +54,8 @@ SNDIO_LIB_DEPENDS.sndio
 # They could just be typos or Portclippy needs to be made aware of them.
 # Please double check them.
 #
-# Prefix them with an _ to tell Portclippy to ignore them.
+# Prefix them with an _ or wrap in '.ifnmake portclippy' to tell
+# Portclippy to ignore them.
 #
 # If in doubt please report this on portfmt's bug tracker:
 # https://github.com/t6/portfmt/issues