+++ tests/clippy/0015.t
@@ -19,7 +19,8 @@ PGSQL_SERVER_VARS
 # They could just be typos or Portclippy needs to be made aware of them.
 # Please double check them.
 #
-# Prefix them with an _ to tell Portclippy to ignore them.
+# Prefix them with an _ or wrap in '.ifnmake portclippy' to tell
+# Portclippy to ignore them.
 #
 # If in doubt please report this on portfmt's bug tracker:
 # https://github.com/t6/portfmt/issues