+++ tests/clippy/0016.t
@@ -8,7 +8,8 @@ license=	GPLv2
 # They could just be typos or Portclippy needs to be made aware of them.
 # Please double check them.
 #
-# Prefix them with an _ to tell Portclippy to ignore them.
+# Prefix them with an _ or wrap in '.ifnmake portclippy' to tell
+# Portclippy to ignore them.
 #
 # If in doubt please report this on portfmt's bug tracker:
 # https://github.com/t6/portfmt/issues